registry/requirements.txt
flask
pymongo[srv]
cerberus
pyyaml